fix(PlayArea): Play area now renders the correct information without error throwing, preventing rerendering
