+++ src/api/index.js
@@ -134,7 +134,6 @@ export const pausePlayback = async() => {
 export const startResumePlayback = async(context) => {
     try {
         // let queryStr = (context ? `?context_uri=${context}` : '')
-        console.log(context)
         const token = window.localStorage.getItem('authorization')
         const response = await fetch(`https://api.spotify.com/v1/me/player/play`, {
             method: "PUT",
